+++ commands.go
@@ -299,7 +299,7 @@ func initCommands(
 				Version:           Version,
 				VersionPrerelease: VersionPrerelease,
 				Platform:          getproviders.CurrentPlatform,
-				CheckFunc:         commandVersionCheck,
+				//CheckFunc:         commandVersionCheck,
 			}, nil
 		},