+++ internal/terraform/context_plan.go
@@ -286,18 +286,6 @@ func (c *Context) plan(config *configs.Config, prevRunState *states.State, opts 
 
 	plan, walkDiags := c.planWalk(config, prevRunState, opts)
 	diags = diags.Append(walkDiags)
-	if diags.HasErrors() {
-		// Non-nil plan along with errors indicates a non-applyable partial
-		// plan that's only suitable to be shown to the user as extra context
-		// to help understand the errors.
-		return plan, diags
-	}
-
-	// The refreshed state ends up with some placeholder objects in it for
-	// objects pending creation. We only really care about those being in
-	// the working state, since that's what we're going to use when applying,
-	// so we'll prune them all here.
-	plan.PriorState.SyncWrapper().RemovePlannedResourceInstanceObjects()
 
 	return plan, diags
 }
@@ -338,10 +326,6 @@ func (c *Context) refreshOnlyPlan(config *configs.Config, prevRunState *states.S
 			"Terraform generated planned resource changes in a refresh-only plan. This is a bug in Terraform.",
 		))
 	}
-
-	// Prune out any placeholder objects we put in the state to represent
-	// objects that would need to be created.
-	plan.PriorState.SyncWrapper().RemovePlannedResourceInstanceObjects()
 
 	// We don't populate RelevantResources for a refresh-only plan, because
 	// they never have any planned actions and so no resource can ever be
@@ -580,9 +564,13 @@ func (c *Context) planWalk(config *configs.Config, prevRunState *states.State, o
 	// we encountered errors, which we'll return as part of a non-nil plan
 	// so that e.g. the UI can show what was planned so far in case that extra
 	// context helps the user to understand the error messages we're returning.
-
 	prevRunState = walker.PrevRunState.Close()
+
+	// The refreshed state may have data resource objects which were deferred
+	// to apply and cannot be serialized.
+	walker.RefreshState.RemovePlannedResourceInstanceObjects()
 	priorState := walker.RefreshState.Close()
+
 	driftedResources, driftDiags := c.driftedResources(config, prevRunState, priorState, moveResults)
 	diags = diags.Append(driftDiags)
 
@@ -643,6 +631,12 @@ func (c *Context) planGraph(config *configs.Config, prevRunState *states.State, 
 	}
 }
 
+// driftedResources is a best-effort attempt to compare the current and prior
+// state. If we cannot decode the prior state for some reason, this should only
+// return warnings to help the user correlate any missing resources in the
+// report. This is known to happen when targeting a subset of resources,
+// because the excluded instances will have been removed from the plan and
+// not upgraded.
 func (c *Context) driftedResources(config *configs.Config, oldState, newState *states.State, moves refactoring.MoveResults) ([]*plans.ResourceInstanceChangeSrc, tfdiags.Diagnostics) {
 	var diags tfdiags.Diagnostics
 
@@ -690,35 +684,35 @@ func (c *Context) driftedResources(config *configs.Config, oldState, newState *s
 					addr.Resource.Resource.Type,
 				)
 				if schema == nil {
-					// This should never happen, but just in case
-					return nil, diags.Append(tfdiags.Sourceless(
-						tfdiags.Error,
+					diags = diags.Append(tfdiags.Sourceless(
+						tfdiags.Warning,
 						"Missing resource schema from provider",
-						fmt.Sprintf("No resource schema found for %s.", addr.Resource.Resource.Type),
+						fmt.Sprintf("No resource schema found for %s when decoding prior state", addr.Resource.Resource.Type),
 					))
+					continue
 				}
 				ty := schema.ImpliedType()
 
 				oldObj, err := oldIS.Current.Decode(ty)
 				if err != nil {
-					// This should also never happen
-					return nil, diags.Append(tfdiags.Sourceless(
-						tfdiags.Error,
+					diags = diags.Append(tfdiags.Sourceless(
+						tfdiags.Warning,
 						"Failed to decode resource from state",
-						fmt.Sprintf("Error decoding %q from previous state: %s", addr.String(), err),
+						fmt.Sprintf("Error decoding %q from prior state: %s", addr.String(), err),
 					))
+					continue
 				}
 
 				var newObj *states.ResourceInstanceObject
 				if newIS != nil && newIS.Current != nil {
 					newObj, err = newIS.Current.Decode(ty)
 					if err != nil {
-						// This should also never happen
-						return nil, diags.Append(tfdiags.Sourceless(
-							tfdiags.Error,
+						diags = diags.Append(tfdiags.Sourceless(
+							tfdiags.Warning,
 							"Failed to decode resource from state",
 							fmt.Sprintf("Error decoding %q from prior state: %s", addr.String(), err),
 						))
+						continue
 					}
 				}